remove old component and style reference
