--- src/components/IconLibrary-Old/ClickableTile/index.js
@@ -1,3 +1,0 @@
-import ClickTile from './ClickableTile';
-
-export default ClickTile;
