--- src/components/IconLibrary-Old/index.js
@@ -1,3 +1,0 @@
-import IconLibrary from './IconLibrary';
-
-export default IconLibrary;
