+++ src/components/IconLibrary/IconLibrary.module.scss
@@ -1,0 +1,15 @@
+.icon-grid {
+  display: grid;
+  width: 100%;
+  grid-template-columns: repeat(6, 1fr);
+}
+
+.icon-container {
+  width: 100%;
+  height: 100%;
+  background-color: $ui-background;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: $spacing-05;
+}
