+++ src/styles/index.scss
@@ -11,7 +11,6 @@
 @import 'src/components/ComponentStatus/component-status.scss';
 @import 'src/components/GlossaryComponent/GlossaryList/glossary-list.scss';
 @import 'src/components/GlossaryComponent/GlossaryNav/glossary-nav.scss';
-@import 'src/components/IconLibrary/icon-library.scss';
 @import 'src/components/TypeScaleTable/type-scale-table.scss';
 @import 'src/components/TypeWeight/type-weight.scss';
 @import 'src/components/HomepageVideo/homepage-video.scss';
